Conflicts:
	rts/Game/GameServer.cpp
	rts/Sim/Path/QTPFS/PathDefines.hpp
	rts/Sim/Path/QTPFS/PathManager.hpp

+++ rts/System/Platform/errorhandler.cpp
@@ -45,10 +45,10 @@ static void ExitMessage(const std::string& msg, const std::string& caption, unsi
 	}
 
 #ifdef _MSC_VER
-	TerminateProcess(GetCurrentProcess(), -1);
-#else
-	exit(-1);
+	if (forced)
+		TerminateProcess(GetCurrentProcess(), -1);
 #endif
+	exit(-1);
 }
